Rework parser.
