+++ src/components/documents/NetpbmEditor/types.tsx
@@ -1,3 +1,11 @@
+export interface RasterParserInput {
+    format: 'P1' | 'P2' | 'P3';
+    width: number;
+    height: number;
+    maxValue?: number;
+    raster: string;
+}
+
 export type ParserMessage = string | React.ReactElement;
 
 export interface ParserImageDataResult {
